+++ packages/mobile/src/hooks/useQuote.ts
@@ -2,20 +2,21 @@ import * as React from 'react';
 import { useNetInfo } from '@react-native-community/netinfo';
 
 export function useQuote() {
-  const netInfo = useNetInfo()
+  const { isConnected } = useNetInfo()
   const [quote, setQuote] = React.useState('Intense and relaxed')
   React.useEffect(() => {
-    if (netInfo.isConnected) {
+    if (isConnected) {
       fetch('https://quotes.rest/qod.json')
         .then(response => response.json())
         .then(responseJson => {
+          console.log("fetching quote", responseJson)
           setQuote(responseJson.contents.quotes[0].quote);
         })
         .catch(error => {
           console.error('QuoteContainer', error);
         });
     }
-  }, []);
+  }, [isConnected]);
 
   return quote;
 }