+++ packages/mobile/wdyr.js
@@ -4,7 +4,7 @@ if (process.env.NODE_ENV === 'development') {
 	const whyDidYouRender = require('@welldone-software/why-did-you-render');
 	const ReactRedux = require('react-redux');
 	whyDidYouRender(React, {
-		trackAllPureComponents: true,
+		trackAllPureComponents: false,
 		trackExtraHooks: [[ReactRedux, 'useSelector']],
 	});
 }